Upgrade to PHP CS Fixer version 2, since 1 has been abandoned for ages
We couldn't just upgrade CS Fixer in the project's composer.json because
CS Fixer depends on a part of Symfony, but we can install CS Fixer into
the hook's private environment, which is therefore what we do now.

It's a little messy because pre-commit doesn't natively support
Composer, but it works fine. :D